Move drone/docker/mod.rs test into an integration test
